Add `testtools` package and build tag
This adds common build tools we can use in our test to a dedicated
directory for cleanliness. They may be enabled via the `testtools` build
tag. This ensures we don't pollute the main binary with things we don't
use.
